requirements.txt
Django==1.6.5
Pillow==2.4.0
Pygments==1.6
South==0.8.4
argparse==1.2.1
bpython==0.13
django-mathfilters==0.2.1
django-suit==0.2.8
djangorestframework==2.3.14
html5lib==0.999
reportlab==2.4
six==1.7.2
wsgiref==0.1.2
